signTxOutput: update for stax
